# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/task": {
            "post": {
                "tags": [
                    "Task"
                ],
                "summary": "Create new task",
                "operationId": "createTask",
                "requestBody": {
                    "$ref": "#/components/requestBodies/TaskRequest"
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Unprocessable Content"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/task/{task_id}/edit": {
            "post": {
                "tags": [
                    "Task"
                ],
                "summary": "Edit task",
                "operationId": "editTask",
                "parameters": [
                    {
                        "name": "task_id",
                        "in": "path",
                        "required": true,
                        "example": "1"
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/TaskEditRequest"
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Unprocessable Content"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/task/list/{status}/all": {
            "get": {
                "tags": [
                    "Task"
                ],
                "summary": "All tasks list by status",
                "operationId": "getAllTasksByStatus",
                "parameters": [
                    {
                        "name": "status",
                        "in": "path",
                        "required": true,
                        "example": "waiting"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/task/{id}/change-status": {
            "post": {
                "tags": [
                    "Task"
                ],
                "summary": "Change task status",
                "operationId": "changeStatus",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "example": "1"
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/TaskStatusRequest"
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Unprocessable Content"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/task/{id}/close": {
            "post": {
                "tags": [
                    "Task"
                ],
                "summary": "Close task",
                "operationId": "closeTask",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Unprocessable Content"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/task/list/{category_id}": {
            "get": {
                "tags": [
                    "Task"
                ],
                "summary": "All tasks by category id",
                "operationId": "getTasksByCategoryId",
                "parameters": [
                    {
                        "name": "category_id",
                        "in": "path",
                        "required": true,
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/task/{id}": {
            "get": {
                "tags": [
                    "Task"
                ],
                "summary": "Task by id",
                "operationId": "getTaskById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/task/{id}/visitor": {
            "get": {
                "tags": [
                    "Task"
                ],
                "summary": "Task by id for auth users",
                "operationId": "getTaskByIdForVisitor",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/task/offer/{id}/change-status": {
            "post": {
                "tags": [
                    "Task"
                ],
                "summary": "Change offer status",
                "operationId": "changeOfferStatus",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "example": "1"
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/OfferStatusRequest"
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Unprocessable Content"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/task/offer/{id}/accept": {
            "post": {
                "tags": [
                    "Task"
                ],
                "summary": "Accept offer",
                "operationId": "acceptOffer",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Unprocessable Content"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/task/offer/{id}/reject": {
            "post": {
                "tags": [
                    "Task"
                ],
                "summary": "Reject offer",
                "operationId": "rejectOffer",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Unprocessable Content"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "requestBodies": {
...
            "OfferStatusRequest": {
                "description": "Change offer status.",
                "required": true,
                "content": {
                    "application/x-www-form-urlencoded": {
                        "schema": {
                            "required": [
                                "status"
                            ],
                            "properties": {
                                "status": {
                                    "type": "string",
                                    "enum": [
                                        "new",
                                        "accepted",
                                        "rejected"
                                    ],
                                    "example": "accepted"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "TaskEditRequest": {
                "description": "Edit task.",
                "required": true,
                "content": {
                    "multipart/form-data": {
                        "schema": {
                            "required": [
                                "name",
                                "description",
                                "status"
                            ],
                            "properties": {
                                "logo": {
                                    "description": "New task logo in base64 format.",
                                    "type": "string"
                                },
                                "name": {
                                    "description": "The task name.",
                                    "type": "string",
                                    "example": "test"
                                },
                                "description": {
                                    "description": "The task description.",
                                    "type": "string",
                                    "example": "test description"
                                },
                                "category_id": {
                                    "description": "The task category id.",
                                    "type": "integer",
                                    "example": "1"
                                },
                                "status": {
                                    "type": "string",
                                    "enum": [
                                        "waiting",
                                        "in progress",
                                        "done",
                                        "closed",
                                        "hidden",
                                        "archived"
                                    ],
                                    "example": "closed"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "TaskRequest": {
                "description": "Create new task.",
                "required": true,
                "content": {
                    "multipart/form-data": {
                        "schema": {
                            "required": [
                                "name",
                                "description"
                            ],
                            "properties": {
                                "logo": {
                                    "description": "New task logo in base64 format.",
                                    "type": "string"
                                },
                                "name": {
                                    "description": "The task name.",
                                    "type": "string",
                                    "example": "test"
                                },
                                "description": {
                                    "description": "The task description.",
                                    "type": "string",
                                    "example": "test description"
                                },
                                "category_id": {
                                    "description": "The task category id.",
                                    "type": "integer",
                                    "example": "1"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "TaskStatusRequest": {
                "description": "Change task status.",
                "required": true,
                "content": {
                    "application/x-www-form-urlencoded": {
                        "schema": {
                            "required": [
                                "status"
                            ],
                            "properties": {
                                "status": {
                                    "type": "string",
                                    "enum": [
                                        "waiting",
                                        "in progress",
                                        "done",
                                        "closed",
                                        "hidden",
                                        "archived"
                                    ],
                                    "example": "closed"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}